Virtual Mouse using OpenCv/requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
PyAutoGUI>=0.9.54
pillow>=10.0.0
keyboard>=0.13.5